+++ Tests/govuk_ios/govuk_ios_snapshot_tests/Helpers/SnapshotTestCase.swift
@@ -13,7 +13,7 @@ class SnapshotTestCase: FBSnapshotTestCase {
         Container.shared.lottieConfiguration.register {
             LottieConfiguration(renderingEngine: .mainThread)
         }
-//        self.recordMode = true
+        // self.recordMode = true
     }
 
     func VerifySnapshotInWindow(_ viewController: UIViewController,
